requirements.txt
albumentations==1.3.1
alembic==1.11.1
altair==4.2.0
altgraph==0.17.3
amqp==5.1.1
anyio==3.6.2
apache-superset==2.1.0
apispec==5.2.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
aspose-words==23.5.0
asttokens==2.0.5
async-generator==1.10
attr==0.3.1
attrs==23.1.0
auto-py-to-exe==2.24.1
azure-core==1.27.0
azure-storage-blob==12.16.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.0
beautifulsoup4==4.11.1
billiard==4.1.0
bleach==3.3.1
blinker==1.6.2
blis==0.7.9
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
bottle==0.12.23
bottle-websocket==0.2.9
boxing==0.1.4
Brotli==1.0.9
cachelib==0.4.1
cachetools==5.3.1
catalogue==2.0.8
celery==5.3.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
chromedriver-binary==110.0.5481.30.0
chromedriver-binary-auto==0.2.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
confection==0.0.4
contextvars==2.4
contourpy==1.0.7
convertdate==2.4.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.6
Django==3.2.19
django-annoying==0.10.6
django-cleanup==7.0.0
django-cors-headers==3.6.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.1
django-dropbox-storage==0.4.1
django-environ==0.9.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.2.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
dropbox==11.36.0
easydict==1.10
Eel==0.14.0
email-validator==1.3.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==0.8.3
expiringdict==1.2.2
fastjsonschema==2.15.3
filelock==3.12.0
Flask==2.1.3
Flask-AppBuilder==4.3.2
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
flatbuffers==23.5.26
fonttools==4.39.2
func-timeout==4.3.5
future==0.18.2
geographiclib==2.0
geopy==2.3.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.27
google-api-core==2.11.0
google-auth==2.14.1
google-auth-oauthlib==0.8.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphlib-backport==1.0.3
greenlet==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.54.2
grpcio-status==1.48.2
gspread==5.7.2
h11==0.14.0
h2==4.1.0
hashids==1.3.1
hijri-converter==2.3.1
holidays==0.17.2
hpack==4.0.0
htmlmin==0.1.12
htmltools==0.1.2
humanfriendly==10.0
humanize==4.6.0
hyperframe==6.0.1
idna==3.4
ijson==3.2.0.post0
imageio==2.31.1
immutables==0.19
importlib-metadata==4.11.3
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==1.1.1
interchange==2021.0.4
ipykernel==6.23.2
ipython==8.3.0
ipython-cypher==0.2.6
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaitaistruct==0.10
kiwisolver==1.4.4
kombu==5.3.0
korean-lunar-calendar==0.3.1
label-studio==1.8.0
label-studio-converter==0.0.53
label-studio-tools==0.0.2
langcodes==3.3.0
langdetect==1.0.9
launchdarkly-server-sdk==7.5.0
lazy_loader==0.2
ldap3==2.9.1
limits==3.5.0
linkify-it-py==2.0.0
llvmlite==0.40.1
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==0.8.4
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
neo4j==5.8.1
neo4j-driver==5.0.1
neo4jrestclient==2.1.1
nest-asyncio==1.5.5
networkx==3.1
nltk==3.6.7
notebook==6.4.11
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.0
openapi-codec==1.3.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.0.10
ordered-set==4.0.2
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
pansi==2020.7.3
paramiko==3.2.0
parsedatetime==2.6
parso==0.8.3
passlib==1.7.4
pathy==0.10.1
pefile==2022.5.30
pgsanity==0.2.9
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
pluggy==1.0.0
plumbum==1.8.2
ply==3.11
polyline==2.0.0
pooch==1.7.0
preshed==3.0.8
prettytable==3.7.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.0
psycopg2-binary==2.9.6
publicsuffix2==2.20191221
pure-eval==0.2.2
py==1.11.0
py2exe==0.13.0.0
py2neo==2021.2.3
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.10.4
pydeck==0.7.1
pydivert==2.1.0
Pygments==2.15.1
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.5.0
PyMatting==1.1.8
PyMeeus==0.5.12
pymongo==4.1.1
Pympler==1.0.1
PyMySQL==1.0.3
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
python-environ==0.4.54
python-geohash==0.8.5
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
redis==3.5.3
regex==2022.4.24
rembg==2.0.49
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-image==0.21.0
scikit-learn==1.1.0
scipy==1.8.0
seaborn==0.12.2
selenium==4.8.0
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.25.1
shiny==0.2.9
shinywidgets==0.1.4
simplejson==3.19.1
six==1.16.0
skillNer==1.0.3
slack-sdk==3.21.3
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.4
srsly==2.4.5
sshtunnel==0.4.0
stack-data==0.2.0
starlette==0.25.0
stone==3.3.1
streamlit==1.9.0
sympy==1.12
tabulate==0.8.10
tenacity==8.2.2
terminado==0.13.3
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
ua-parser==0.16.1
uc-micro-py==1.0.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==2.0.2
urwid==2.1.2
user-agents==2.2.0
uvicorn==0.20.0
validators==0.20.0
vine==5.0.0
virtualenv==20.23.0
waitress==2.1.2
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
whichcraft==0.6.1
whitenoise==6.4.0
widgetsnbextension==3.6.0
wordcloud==1.8.2.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==2.3.3
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
xmljson==0.2.0
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.5.2
zstandard==0.19.0
